refactor(s3): move to traits
